+++ ios/Classes/SwiftFlutterVoipKitPlugin.swift
@@ -56,6 +56,14 @@ public class SwiftFlutterVoipKitPlugin: NSObject, FlutterPlugin {
         registrar.addMethodCallDelegate(instance, channel: methodChannel)
     }
     
+    ///useful for integrating with VIOP notifications
+    static public func reportIncomingCall(handle: String, uuid: String, result: FlutterResult?){
+        SwiftFlutterVoipKitPlugin.callController.reportIncomingCall(uuid: UUID(uuidString: uuid)!, handle: handle) { (error) in
+            print("ERROR: \(error?.localizedDescription ?? "none")")
+            result?(error == nil)
+        }
+    }
+    
     //TODO: remove these defaults and get as arguments
     public func handle(_ call: FlutterMethodCall, result: @escaping FlutterResult) {
         let args = call.arguments as? Dictionary<String, Any>
@@ -69,10 +77,7 @@ public class SwiftFlutterVoipKitPlugin: NSObject, FlutterPlugin {
             }
         }else if call.method == SwiftFlutterVoipKitPlugin._methodChannelReportIncomingCall{
             if let handle = args?["handle"] as? String, let uuid = args?["uuid"] as? String{
-                SwiftFlutterVoipKitPlugin.callController.reportIncomingCall(uuid: UUID(uuidString: uuid)!, handle: handle) { (error) in
-                    print("ERROR: \(error?.localizedDescription ?? "none")")
-                    result(error == nil)
-                }
+                SwiftFlutterVoipKitPlugin.reportIncomingCall(handle: handle, uuid: uuid, result: result)
             }else{
                 result(FlutterError.init(code: "bad args", message: nil, details: nil))
             }